+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.70.1
 
+- CDSI enclave updated for production.